Created test_lib.scylla_bench_tools and create_scylla_bench_table_query()
